github_issue_lister_agent/requirements.txt
PyGithub==2.6.1
google-adk==1.6.1
mcp==1.12.0
google-cloud-aiplatform[adk,agent_engines]
cloudpickle